reorg sessions, improve session runner usability
